Use f32, exclusively

+++ src/aabb.rs
@@ -23,51 +23,51 @@ impl AABB {
         self.expand_by_point(&expansion_box.maximum);
     }
 
-    pub fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> bool {
+    pub fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> bool {
 
         // X
-        let inverse_d : f64 = 1.0 / ray.direction.x;
-        let mut t_0 : f64 = (self.minimum.x - ray.origin.x) * inverse_d; 
-        let mut t_1 : f64 = (self.maximum.x - ray.origin.x) * inverse_d;
+        let inverse_d : f32 = 1.0 / ray.direction.x;
+        let mut t_0 : f32 = (self.minimum.x - ray.origin.x) * inverse_d; 
+        let mut t_1 : f32 = (self.maximum.x - ray.origin.x) * inverse_d;
 
         if inverse_d < 0.0 {
             std::mem::swap(&mut t_0, &mut t_1);
         }
 
-        let t_min : f64 = if t_0 > t_min { t_0 } else { t_min };
-        let t_max : f64 = if t_1 < t_max { t_1 } else { t_max };
+        let t_min : f32 = if t_0 > t_min { t_0 } else { t_min };
+        let t_max : f32 = if t_1 < t_max { t_1 } else { t_max };
                         
         if t_max <= t_min { return false; }
 
 
 
         // Y
-        let inverse_d : f64 = 1.0 / ray.direction.y;
-        let mut t_0 : f64 = (self.minimum.y - ray.origin.y) * inverse_d; 
-        let mut t_1 : f64 = (self.maximum.y - ray.origin.y) * inverse_d;
+        let inverse_d : f32 = 1.0 / ray.direction.y;
+        let mut t_0 : f32 = (self.minimum.y - ray.origin.y) * inverse_d; 
+        let mut t_1 : f32 = (self.maximum.y - ray.origin.y) * inverse_d;
 
         if inverse_d < 0.0 {
             std::mem::swap(&mut t_0, &mut t_1);
         }
 
-        let t_min : f64 = if t_0 > t_min { t_0 } else { t_min };
-        let t_max : f64 = if t_1 < t_max { t_1 } else { t_max };
+        let t_min : f32 = if t_0 > t_min { t_0 } else { t_min };
+        let t_max : f32 = if t_1 < t_max { t_1 } else { t_max };
                         
         if t_max <= t_min { return false; }
 
 
 
         // Z
-        let inverse_d : f64 = 1.0 / ray.direction.z;
-        let mut t_0 : f64 = (self.minimum.z - ray.origin.z) * inverse_d; 
-        let mut t_1 : f64 = (self.maximum.z - ray.origin.z) * inverse_d;
+        let inverse_d : f32 = 1.0 / ray.direction.z;
+        let mut t_0 : f32 = (self.minimum.z - ray.origin.z) * inverse_d; 
+        let mut t_1 : f32 = (self.maximum.z - ray.origin.z) * inverse_d;
 
         if inverse_d < 0.0 {
             std::mem::swap(&mut t_0, &mut t_1);
         }
 
-        let t_min : f64 = if t_0 > t_min { t_0 } else { t_min };
-        let t_max : f64 = if t_1 < t_max { t_1 } else { t_max };
+        let t_min : f32 = if t_0 > t_min { t_0 } else { t_min };
+        let t_max : f32 = if t_1 < t_max { t_1 } else { t_max };
                         
         if t_max <= t_min { return false; }
